refactor: rename package from wasseswidgets to widgets
Update applicationId, namespace, all package declarations, imports,
directory structure, and display name from "Wasses Widgets" to "Widgets".
