--- app/src/main/java/com/wassupluke/wasseswidgets/data/api/WeatherApiModels.kt
@@ -1,29 +1,0 @@
-package com.wassupluke.wasseswidgets.data.api
-
-import kotlinx.serialization.SerialName
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class WeatherResponse(
-    val current: CurrentWeather
-)
-
-@Serializable
-data class CurrentWeather(
-    val time: String,
-    @SerialName("temperature_2m") val temperatureCelsius: Float
-)
-
-@Serializable
-data class GeocodingResponse(
-    val results: List<GeocodingResult>? = null
-)
-
-@Serializable
-data class GeocodingResult(
-    val name: String,
-    val latitude: Double,
-    val longitude: Double,
-    val country: String? = null,
-    @SerialName("admin1") val state: String? = null
-)
