--- app/src/main/java/com/wassupluke/wasseswidgets/widget/WeatherWidgetReceiver.kt
@@ -1,8 +1,0 @@
-package com.wassupluke.wasseswidgets.widget
-
-import androidx.glance.appwidget.GlanceAppWidget
-import androidx.glance.appwidget.GlanceAppWidgetReceiver
-
-class WeatherWidgetReceiver : GlanceAppWidgetReceiver() {
-    override val glanceAppWidget: GlanceAppWidget = WeatherWidget()
-}
