+++ app/src/main/java/com/wassupluke/widgets/widget/WeatherWidgetReceiver.kt
@@ -1,0 +1,8 @@
+package com.wassupluke.widgets.widget
+
+import androidx.glance.appwidget.GlanceAppWidget
+import androidx.glance.appwidget.GlanceAppWidgetReceiver
+
+class WeatherWidgetReceiver : GlanceAppWidgetReceiver() {
+    override val glanceAppWidget: GlanceAppWidget = WeatherWidget()
+}
